refactor: moved form file to ui/elements
